Bullet-proofed contribflow to handle 404, 503 and 504 HTTP error codes (not found, service unavailable and timeout)
Also cleaned up the files in lib and test for JSHint.

Added files in bin for JSHint coverage.

Testing:
Grunt runs the tests now :)
Manual testing: Create and delete a feature branch.
Tested submitting against the video-js-swf repo, which still returns a 404
(but contribflow now returns a message to the user that they should try to submit the PR manually).
Submitting this branch as a pull request will be a test of that part of the code.